Fix filtertag component
